requirements.txt
# python_version >= 3.4
pyserial >= 3.5